take package `batch` apart
moved some parts of it to kv. KeyRange is now keys.Range,
Short() is now String() on *BatchRequest in proto (changed so
that it doesn't import 'keys' any more), the rest mostly went
to 'client'.
Did a bunch of gorename (to unexport a few things). No change
of semantics in this commit.
