+++ kv/local_sender.go
@@ -23,7 +23,6 @@ import (
 
 	"golang.org/x/net/context"
 
-	"github.com/cockroachdb/cockroach/batch"
 	"github.com/cockroachdb/cockroach/client"
 	"github.com/cockroachdb/cockroach/proto"
 	"github.com/cockroachdb/cockroach/storage"
@@ -39,7 +38,7 @@ type LocalSender struct {
 }
 
 var _ client.Sender = &LocalSender{}
-var _ batch.Sender = &LocalSender{}
+var _ client.BatchSender = &LocalSender{}
 var _ rangeDescriptorDB = &LocalSender{}
 
 // NewLocalSender returns a local-only sender which directly accesses
@@ -181,7 +180,7 @@ func (ls *LocalSender) SendBatch(ctx context.Context, ba *proto.BatchRequest) (*
 // the command is being executed locally, and the replica is
 // determined via lookup through each store's LookupRange method.
 func (ls *LocalSender) Send(ctx context.Context, call proto.Call) {
-	batch.SendCallConverted(ls, ctx, call)
+	client.SendCallConverted(ls, ctx, call)
 	return
 }
 
@@ -239,7 +238,7 @@ func (ls *LocalSender) firstRange() (*proto.RangeDescriptor, error) {
 // rangeLookup implements the rangeDescriptorDB interface. It looks up
 // the descriptors for the given (meta) key.
 func (ls *LocalSender) rangeLookup(key proto.Key, options lookupOptions, _ *proto.RangeDescriptor) ([]proto.RangeDescriptor, error) {
-	ba, unwrap := batch.MaybeWrap(&proto.RangeLookupRequest{
+	ba, unwrap := client.MaybeWrap(&proto.RangeLookupRequest{
 		RequestHeader: proto.RequestHeader{
 			Key:             key,
 			ReadConsistency: proto.INCONSISTENT,
